Renamed all variables/fields named pathname to dashboard_id
I was previously using the pathname of the admin urls as a unique identifier
for the dashboards. So I introduced the DashboardPreference field "pathname",
and used that word too for misc related variables in the code. But as Mikhail
Korobov suggested it[1] it makes more sense to use the dashboard.get_id()
method instead of computing a uid from some request attribute. This is what I
did in my previous commit. So to be consistant I renamed in this commit
everything named "pathname" to "dashboard_id" in the code.